Reestruturado projeto para colocar todos os ids na pasta entity/id
