+++ api/src/main/java/com/danceSchool/api/entity/id/InstrutorModalidadeId.java
@@ -1,8 +1,9 @@
-package com.danceSchool.api.entity.salaModalidadeId;
+package com.danceSchool.api.entity.id;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.Hibernate;
@@ -14,13 +15,14 @@ import java.util.Objects;
 @Getter
 @Setter
 @Embeddable
-public class SalaModalidadeId implements Serializable {
+public class InstrutorModalidadeId implements Serializable {
     @Serial
-    private static final long serialVersionUID = -7847763584011370940L;
+    private static final long serialVersionUID = 5969865838755503744L;
 
+    @Size(max = 11)
     @NotNull
-    @Column(name = "id_sala", nullable = false)
-    private Integer idSala;
+    @Column(name = "cpf_instrutor", nullable = false, length = 11)
+    private String cpfInstrutor;
 
     @NotNull
     @Column(name = "id_modalidade", nullable = false)
@@ -30,14 +32,14 @@ public class SalaModalidadeId implements Serializable {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        SalaModalidadeId entity = (SalaModalidadeId) o;
-        return Objects.equals(this.idSala, entity.idSala) &&
+        InstrutorModalidadeId entity = (InstrutorModalidadeId) o;
+        return Objects.equals(this.cpfInstrutor, entity.cpfInstrutor) &&
                 Objects.equals(this.idModalidade, entity.idModalidade);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(idSala, idModalidade);
+        return Objects.hash(cpfInstrutor, idModalidade);
     }
 
 }
