+++ api/src/main/java/com/danceSchool/api/entity/instrutorModalidade/InstrutorModalidadeRepository.java
@@ -1,6 +1,6 @@
 package com.danceSchool.api.entity.instrutorModalidade;
 
-import com.danceSchool.api.entity.instrutorModalidadeId.InstrutorModalidadeId;
+import com.danceSchool.api.entity.id.InstrutorModalidadeId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InstrutorModalidadeRepository extends JpaRepository<InstrutorModalidade, InstrutorModalidadeId> {
